refactor: split dxf_builder.py into focused sub-modules (closes #58)
Extract six single-responsibility sub-modules from the 651-line
dxf_builder.py — dxf_math, dxf_text, dxf_layer, dxf_symbols,
dxf_geometry, dxf_arrows.  dxf_builder.py becomes a thin backward-
compatible facade that re-exports every public name unchanged.  Add
54 new direct-import unit tests (one test class per sub-module).
Bump SPEC.md to 0.1.1 and update module descriptions.
